refactor: convert main-component to home-page-component
