+++ src/App.tsx
@@ -1,13 +1,13 @@
 import Footer from './components/footer/footer.tsx';
 import Header from './components/header/header.tsx';
-import Main from './components/main/main-component.tsx';
+import Home from './components/pages/home-page/home-page-component.tsx';
 import './App.css';
 
 function App() {
   return (
     <>
       <Header />
-      <Main />
+      <Home />
       <Footer />
     </>
   );
