+++ src/components/pages/home-page/results-section/error-button/error-button.tsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import '../../main-component.css';
+import '../../home-page.css';
 
 export default function ErrorButton() {
   const [stateChanged, changeState] = useState(0);
