+++ src/components/pages/home-page/results-section/message-components/nothing-message.tsx
