+++ src/components/pages/home-page/results-section/results-section.css
