+++ src/test/nothing-message.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import NothingMessage from '../components/main/results-section/message-components/nothing-message.tsx';
+import NothingMessage from '../components/pages/home-page/results-section/message-components/nothing-message.tsx';
 
 describe('NothingMessage component', () => {
   it('renders correctly', () => {
