+++ src/test/results-section.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import Results from '../components/main/results-section/results-section.tsx';
+import Results from '../components/pages/home-page/results-section/results-section.tsx';
 
 describe('Results component', () => {
   it('renders correctly', () => {
